+++ security/provenance/type.c
@@ -82,6 +82,8 @@ static const char RL_STR_CLOSED[]                = "closed";
 static const char RL_STR_ARG[]                   = "arg";
 static const char RL_STR_ENV[]                   = "env";
 static const char RL_STR_LOG[]                   = "log";
+static const char RL_STR_SH_ATTACH_READ[]        = "sh_attach_read";
+static const char RL_STR_SH_ATTACH_WRITE[]       = "sh_attach_write";
 
 /* node string name */
 static const char ND_STR_UNKNOWN[]                           = "unknown";
@@ -251,6 +253,10 @@ const char* relation_str(uint64_t type)
 		return RL_STR_ENV;
 	case RL_LOG:
 		return RL_STR_LOG;
+	case RL_SH_ATTACH_READ:
+		return RL_STR_SH_ATTACH_READ;
+	case RL_SH_ATTACH_WRITE:
+		return RL_STR_SH_ATTACH_WRITE;
 	default:
 		return RL_STR_UNKNOWN;
 	}
@@ -327,6 +333,8 @@ uint64_t relation_id(const char* str)
 	MATCH_AND_RETURN(str, RL_STR_ARG, RL_ARG);
 	MATCH_AND_RETURN(str, RL_STR_ENV, RL_ENV);
 	MATCH_AND_RETURN(str, RL_STR_LOG, RL_LOG);
+	MATCH_AND_RETURN(str, RL_STR_SH_ATTACH_READ, RL_SH_ATTACH_READ);
+	MATCH_AND_RETURN(str, RL_STR_SH_ATTACH_WRITE, RL_SH_ATTACH_WRITE);
 	return 0;
 }
 EXPORT_SYMBOL_GPL(relation_id);